+++ Business/Concretes/ApplicationStateManager.cs
@@ -5,7 +5,7 @@ using Business.Responses.ApplicationStates;
 using Core.Utilities.Results;
 using DataAccess.Abstracts;
 using Entities.Concretes;
-using static System.Net.Mime.MediaTypeNames;
+using Microsoft.EntityFrameworkCore;
 
 namespace Business.Concretes
 {
@@ -14,49 +14,81 @@ namespace Business.Concretes
         private readonly IApplicationStateRepository _applicationStateRepository;
         private readonly IMapper _mapper;
 
-        public ApplicationStateManager(IApplicationStateRepository applicaitonStateRepository, IMapper mapper)
+        public ApplicationStateManager(IApplicationStateRepository applicationstateRepository, IMapper mapper)
         {
-            _applicationStateRepository = applicaitonStateRepository;
+            _applicationStateRepository = applicationstateRepository;
             _mapper = mapper;
         }
 
         public async Task<IDataResult<CreateApplicationStateResponse>> AddAsync(CreateApplicationStateRequest request)
         {
             ApplicationState applicationState = _mapper.Map<ApplicationState>(request);
+
             await _applicationStateRepository.AddAsync(applicationState);
+
             CreateApplicationStateResponse response = _mapper.Map<CreateApplicationStateResponse>(applicationState);
-            return new SuccessDataResult<CreateApplicationStateResponse>(response, "Ekleme işlemi başarılı");
+
+            return new SuccessDataResult<CreateApplicationStateResponse>(response, "Added Successfully");
         }
 
         public async Task<IDataResult<DeleteApplicationStateResponse>> DeleteAsync(DeleteApplicationStateRequest request)
         {
-            ApplicationState applicationState = _mapper.Map<ApplicationState>(request);
+            var applicationState = await _applicationStateRepository.GetByIdAsync(predicate: applicationState => applicationState.Id == request.Id);
+
+            if (applicationState == null)
+            {
+                return new ErrorDataResult<DeleteApplicationStateResponse>("ApplicationState not found");
+            }
+
             await _applicationStateRepository.DeleteAsync(applicationState);
-            DeleteApplicationStateResponse response = _mapper.Map<DeleteApplicationStateResponse>(applicationState);
-            return new SuccessDataResult<DeleteApplicationStateResponse>(response, "Silme işlemi başarılı");
+
+            var response = _mapper.Map<DeleteApplicationStateResponse>(applicationState);
+
+            return new SuccessDataResult<DeleteApplicationStateResponse>(response, "Deleted Successfully");
         }
+
 
         public async Task<IDataResult<List<GetAllApplicationStateResponse>>> GetAllAsync()
         {
-            List<ApplicationState> applicationStates = await _applicationStateRepository.GetAllAsync();
+            List<ApplicationState> applicationStates = await _applicationStateRepository.GetAllAsync(include: x => x.Include(x => x.Application).Include(x => x.Application.Bootcamp));
+            
             List<GetAllApplicationStateResponse> responses = _mapper.Map<List<GetAllApplicationStateResponse>>(applicationStates);
-            return new SuccessDataResult<List<GetAllApplicationStateResponse>>(responses, "Veriler başarıyla listelendi.");
+            
+            return new SuccessDataResult<List<GetAllApplicationStateResponse>>(responses, "Listed Successfully");
         }
 
-        public async Task<IDataResult<GetByIdApplicationStateResponse>> GetByIdAsync(int id)
+        public async Task<IDataResult<GetByIdApplicationStateResponse>> GetByIdAsync(GetByIdApplicationStateRequest request)
         {
-            ApplicationState applicationState = await _applicationStateRepository.GetAsync(x => x.Id == id);
-            GetByIdApplicationStateResponse response = _mapper.Map<GetByIdApplicationStateResponse>(applicationState);
-            return new SuccessDataResult<GetByIdApplicationStateResponse>(response);
+            var applicationState = await _applicationStateRepository.GetByIdAsync(predicate: applicationState => applicationState.Id == request.Id);
+
+            if (applicationState == null)
+            {
+                return new ErrorDataResult<GetByIdApplicationStateResponse>("ApplicationState not found");
+            }
+
+            await _applicationStateRepository.DeleteAsync(applicationState);
+
+            var response = _mapper.Map<GetByIdApplicationStateResponse>(applicationState);
+
+            return new SuccessDataResult<GetByIdApplicationStateResponse>(response, "Showed Successfully");
         }
 
         public async Task<IDataResult<UpdateApplicationStateResponse>> UpdateAsync(UpdateApplicationStateRequest request)
         {
-            ApplicationState applicationState = await _applicationStateRepository.GetAsync(x=>x.Id == request.Id);
+            var applicationState = await _applicationStateRepository.GetByIdAsync(predicate: applicationState => applicationState.Id == request.Id);
+
+            if (applicationState == null)
+            {
+                return new ErrorDataResult<UpdateApplicationStateResponse>("ApplicationState not found");
+            }
+
             _mapper.Map(request, applicationState);
+
             await _applicationStateRepository.UpdateAsync(applicationState);
-            UpdateApplicationStateResponse response = _mapper.Map<UpdateApplicationStateResponse>(applicationState);
-            return new SuccessDataResult<UpdateApplicationStateResponse>(response, "Güncelleme işlemi başarılı");
+
+            var response = _mapper.Map<UpdateApplicationStateResponse>(applicationState);
+
+            return new SuccessDataResult<UpdateApplicationStateResponse>(response, "Updated Successfully");
         }
     }
 }
